requirements.txt
playwright==1.44.0
python-telegram-bot==21.3
requests==2.32.3
pycryptodome==3.20.0